Reduce unnecessary amount of group announcement spam
Instead of every peer refreshing their announcements every n minutes we ask
each peer to play a lottery every n minutes where their chance of winning
is relative to the number of peers in the group, and announce only if they
win. We only need a bare minimum of one announcement in the DHT at any given
time, so this ensures that while not constantly burdening the DHT with
unnecessary traffic.